chore: move mock library to Confirma.Classes.Mock namespace
